Refactor header bidding MoPub adapter names
