+++ build.gradle
@@ -25,7 +25,7 @@ subprojects {
         branchName = "${System.getenv("CIRCLE_BRANCH")}"
         buildNumber = "${System.getenv("CIRCLE_BUILD_NUM")}"
     }
-    version = "0.11.1"
+    version = "0.12.0"
     group = "net.pubnative"
 }
